refactor: RandomNumbersGenerator 및 NumbersGenerator를 domain 패키지로 이동
