+++ src/main/java/domain/NumbersGenerator.java
@@ -1,4 +1,4 @@
-package utils;
+package domain;
 
 import java.util.List;
