+++ src/test/java/domain/LottoStoreTest.java
@@ -2,8 +2,6 @@ package domain;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import utils.NumbersGenerator;
-import utils.RandomNumbersGenerator;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.junit.jupiter.api.Assertions.assertAll;
