+++ src/test/java/domain/RandomNumbersGeneratorTest.java
@@ -1,8 +1,9 @@
-package utils;
+package domain;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
 import java.util.List;
+
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
