+++ src/test/java/domain/WinningResultTest.java
@@ -6,7 +6,6 @@ import java.util.List;
 import java.util.Map;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import utils.NumbersGenerator;
 
 class WinningResultTest {
